Implement savestates
We implement two savestate approaches. One is file-based and one is memory-based which we refer to as a context switch. The file-based approach simply reads and writes inputs to a file, playing the inputs back on load in order to reach the frame of the savestate. This approach is reliable, but does not have optimal save/load times as the filesize scales with framecount. Our memory-based savestate approach is basically able to perform a direct memcpy of the entire heap, since we implement EGG::Heap. This uses a total of 168MB for 10 savestate slots, but we do not initialize this memory unless the slot is used. This approach offers constant-time saves and loads regardless of framecount, but the states do not persist across Kinoko instances.